refactor(Commands): New logic foundation these implemented in ProviderMakeCommand.php and FactoryMakeCommand.php
