+++ src/Contracts/Application.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace KoalaFacade\DiamondConsole\Contracts;
+
+interface Application extends Options, Console, Placeholders, Arguments
+{
+}
