+++ src/Contracts/Arguments.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace KoalaFacade\DiamondConsole\Contracts;
+
+interface Arguments
+{
+    public function resolveNameArgument(): string;
+}
