--- src/Contracts/Middleware.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace KoalaFacade\DiamondConsole\Contracts;
-
-interface Middleware
-{
-    public function resolveMiddleware(): void;
-}
