+++ src/Contracts/Options.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace KoalaFacade\DiamondConsole\Contracts;
+
+interface Options
+{
+    public function resolveForceOption(): bool;
+}
