+++ src/Contracts/Placeholders.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace KoalaFacade\DiamondConsole\Contracts;
+
+use KoalaFacade\DiamondConsole\DataTransferObjects\PlaceholderData;
+
+interface Placeholders
+{
+    public function resolvePlaceholders(): PlaceholderData;
+}
