+++ src/Exceptions/FileAlreadyExistException.php
@@ -9,7 +9,7 @@ class FileAlreadyExistException extends \Exception
     public function __construct(public string $fileName)
     {
         parent::__construct(
-            message: $this->fileName . 'already exists.',
+            message: $this->fileName . ' already exists.',
             code: Response::HTTP_FORBIDDEN,
         );
     }
